The five things CodeQL found, fixed rather than dismissed
An integer multiplication was cast to a float after the fact, so a large enough
length would have overflowed before it ever became a float. It multiplies as a
float from the start now.

Three calls to Path.Combine silently drop everything before a rooted argument,
which turns a path built from parts into a path built from the last part. Path.Join
never does that, and the parts here are exactly the kind that grow over time.

The fifth was the entry point xunit generates into obj/, reported as a style
finding on every run. Build output is not code anyone wrote, so CodeQL no longer
reads it -- a list that always has something in it is a list nobody reads.

Dependabot's security updates and the vulnerability alerts are switched on, so an
advisory in a dependency arrives as a pull request rather than as nothing.

Gate: 37 tests, 0 failures, 0 warnings, 100% of lines and branches.

+++ Tst/DeepSharp/Tensors/CpuBackendTests.cs
@@ -56,7 +56,7 @@ public class CpuBackendTests
         // number of registers is exactly where an off-by-one hides — and it hides quietly, because the first
         // few thousand values are right.
         float[] left = [.. Enumerable.Range(0, length).Select(i => (float)i)];
-        float[] right = [.. Enumerable.Range(0, length).Select(i => (float)(i * 2))];
+        float[] right = [.. Enumerable.Range(0, length).Select(i => i * 2f)];
 
         var sum = _backend.Add(Tensor.From(new Shape(length), left), Tensor.From(new Shape(length), right));
